[GPJ-30] Reutiliza definição de MyErrorStateMatcher
