+++ ProcessosFront/src/components/signup-form/signup-form.component.ts
@@ -1,32 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import {
-  FormControl,
-  FormGroupDirective,
-  NgForm,
-  Validators,
-} from '@angular/forms';
-import { ErrorStateMatcher } from '@angular/material/core';
+import { FormControl, Validators } from '@angular/forms';
 import { nameValidator } from 'src/validators/name';
 import { inputNumberValdiator } from 'src/validators/input-numer';
 import { passwordValidator } from 'src/validators/password';
 import { confPasswordValidator } from 'src/validators/confirm-password';
 import { User } from '../../../../common/user';
 import { UserService } from 'src/services/user.service';
-
-export class MyErrorStateMatcher implements ErrorStateMatcher {
-  isErrorState(
-    control: FormControl | null,
-    form: FormGroupDirective | NgForm | null
-  ): boolean {
-    const isSubmitted = form && form.submitted;
-    return !!(
-      control &&
-      control.invalid &&
-      (control.dirty || control.touched || isSubmitted)
-    );
-  }
-}
+import { MyErrorStateMatcher } from 'src/validators/error-state-matcher';
 
 @Component({
   selector: 'app-signup-form',
